requirements.txt.txt
langgraph>=0.0.26
langchain>=0.1.20
openai>=1.22.0
google-search-results>=2.4.2
pymongo>=4.6.3
pydantic>=2.6.4
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
scikit-learn>=1.4.1.post1
pandas>=2.2.2
numpy>=1.26.4
tqdm>=4.66.4
streamlit>=1.35.0
httpx>=0.27.0
typing-extensions>=4.11.0
